9G$E *4BMED3DYY3H*LJ:5Y97BKJBOGE*4I19RWL-%0KQ2V
5sg6++++++++zspJrKvPBVWxxpBwE65JlW7ZQJQXQ6M-Ov8P1Qq0d3UjJJfw79CaG64YaiViJkBqCykFxaFtYVqGGhkMqRMPqu9pzNpnjgBuOlrDPwKC3qjebSp38k9HQ2frjzx7kwa8bCdvnGGOk6tS92yKp93fSHNvmzJK43I9HMrJqhsKEoz-dbRSygxHr0f7FdaKP2Dbqw5gffEkNP--u57qOQZn7olEBQK29Tgn2ZdH5Hyw0ikdP2IUNR0-e+DGVGpSFFuQGJKnbpZ54qphJyHXZBfbif4Zq2G1Rw9MAxLKSCI1arcgNvAL9yWR5SchmzGjbwo8SUwsPhVvNIpN1wul0JJ8L82Sruozaqp4x6BKfT4tesDdmBXPr5k8xxE95qXtyXI7-oWZR8UWBdd9J5oH-oU7rb6EdFV0t2NKzRM7npJh-lACGhxMajeGRzERDMOYRn3cpuBCPgoCC51Kyk8cFDnE0H8UBTv2PYl+ZrHBEUAPI74fIgfYmEQlojoTTtxwQsNtC02vkOYwojG1MnriAFuAkixCVH3CZcOuteNAwB5lzW0VR11ICTUNc94W2LIInIwrjplGcpX9AxFpLAHAu5gkKGuGtduRtCeXhsMMFKB5IGohCoyCKy4Y-h-FTXWa-A1W8TQS56e+8Es3g3s9yIaMaWj55LQP-fv5pNNp5vzx1hj-Qdv2JuJdgZ9wCYuny9RQaHLfeVHslS4p3ZpTeKNueT6UmSny8zlN3QqzyVb0p0lPzh22Bmucy1u-LomoJs+sNteLNLao3pb8yz1wE3swkKilr4NE4fIQZ-Ty9waHY4PEycg06aZLs1JjA2SqT3lFX0srdQDY9MQe4ZfZAqlMCcSTxAkaYz0Hu40-oP6K7RrYP-RjgiMSF6sJID0X1l1E-vfzvGygBbZ6dlDIeRO7U2f6MtUZmt9SCykH8aDHV2MyCR64oAhMIOqyDOBsZjxLDeqIBLeA0mYz1XtoO1E9x5njUXjiEpnMsTHEd2E61AFxNIopXNaL0bou-gBtXv4meEbc5Pjw8etamkOa4fXOMf1tZ0T49Z8k3lozqfAjuR7CJZNUrl2aQtArePw9pHG8d79aNQ-Rkf1a+5c1P03BPgNPAQtUy+NCwksXHpQ9ZXbv9suag9WFUsA-w7dCHeBZptW12lYnMEmI3k5rBCyI4HkhGReuUC33s1-mUnK3CjnWW41xi1E4AO4M4zsJBWQnNk0V3euBHgfpPYqpNa2S1l2Fe0V+EV3kk+20SvIwWIPiMIt65pgiy0uu7gleA-Akc96FGgBFM2yD+ZqjMo8TjmfPguaWCJRF02Qd+0tZU6W8xtNR5DOlkLZrFp5SOpehXZSfVotcTbKBNh9R9CZfkhBm5xtoATfN0KgB3gObAy87WUiIQT5OxUwVwOuCktZ6Oyi7vrffIKon7ismbKTvunvip5EnyC+4gC-kszw1noUkMbA6+++

[blake2s:base45|gzip:xxencode]

+++ src/surreal_db.rs
@@ -3549,7 +3549,7 @@ impl SurrealDatabase {
             .filter_map(|v| v.as_str().map(|s| s.to_string()))
             .collect();
         let current_index = obj["current_index"].as_u64().unwrap_or(0) as usize;
-        let current_chunk_index = obj["current_chunk_index"].as_u64().unwrap_or(0) as u8;
+        let current_chunk_index = obj["current_chunk_index"].as_u64().unwrap_or(0) as u16;
         let step = obj["step"].as_u64().unwrap_or(0) as u32;
         let attempts_on_current = obj["attempts_on_current"].as_u64().unwrap_or(0) as u8;
         let remembered_count = obj["remembered_count"].as_u64().unwrap_or(0) as u32;
